+++ website/src/components/index.ts
@@ -1,2 +1,1 @@
 export { Footer } from './footer';
-export { Logo } from './logo';
