+++ website/src/components/logos/index.ts
@@ -1,2 +1,3 @@
+export { Logo } from './logo';
 export { LeetCodeLogo } from './leetcode';
 export { ProgrammersLogo } from './programmers';
